+++ CHANGELOG.md
@@ -36,9 +36,12 @@ the detailed section referring to by linking pull requests or issues.
 * SQL-based TransferProcessStore (#865)
 * Add instructions for observability sample with Azure Application Insights (#928)
 * Add interface `WebServer` to `web-spi` (#921)
+* Add MicrometerExtension integration tests (#935)
 * Implement Asset service for Data Management API (#931)
 * Implement ContractDefinition service for Data Management API (#940)
 * Implement ContractNegotiation service for Data Management API (#957)
+* In-memory implementation of PolicyStore (#930)
+* Implement AssetLoader, AssetIndex, DataAddressResolver for SQL (#863)
 
 #### Changed